update to libnm-1-8.28
Also, add "0001-nm-applet-no-notifications.patch". This patch was
originally carried by "network-manager-applet" (which shipped the
gschema before). There is no history as to why this is correct to
do, but assume it is and keep the patch.

Also, for now disable build of gtk4 library. It doesn't seem stable
enough and doesn't build on F32. Reenable it later.

Also, disable build of vapi. For libnm, we ship vapi files in
NetworkManager-libnm-devel package. That doesn't seem right. First
figure out in which libnma package to ship them (if at all).